+++ src/main/java/com/place_application/demo/service/impl/TeacherServiceImpl.java
@@ -1,6 +1,13 @@
 package com.place_application.demo.service.impl;
 
+import com.place_application.demo.config.JwtConfig;
 import com.place_application.demo.dao.TeacherDao;
+import com.place_application.demo.pojo.Admin;
+import com.place_application.demo.pojo.Position;
+import com.place_application.demo.pojo.TeacherQueryAssist;
+import com.place_application.demo.service.PositionService;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.SignatureException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
@@ -8,6 +15,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.place_application.demo.pojo.Teacher;
 import com.place_application.demo.service.TeacherService;
 
+import javax.annotation.Resource;
 import java.util.List;
 
 @Service
@@ -15,6 +23,10 @@ import java.util.List;
 public class TeacherServiceImpl implements TeacherService {
     @Autowired
     private TeacherDao teacherDao;
+    @Autowired
+    private PositionService positionService;
+    @Resource
+    private JwtConfig jwtConfig;
 
     @Override
     public Teacher registerTeacher(Teacher teacher) {
@@ -30,9 +42,30 @@ public class TeacherServiceImpl implements TeacherService {
     public Teacher loginTeacher(Teacher teacher) {
         Teacher teacherTarget = this.teacherDao.selectTeacherByNo(teacher.getTeacher_no());
         if (teacherTarget != null && teacherTarget.getTeacher_password().equals(teacher.getTeacher_password())) {
+            teacherTarget.setTeacher_password("");
             return  teacherTarget;
         }
         return null;
+    }
+
+    @Override
+    public Teacher loginTeacherWithToken(String token) {
+        Claims claims;
+        try{
+            claims = jwtConfig.getTokenClaim(token);
+            if(claims == null || jwtConfig.isTokenExpired(claims.getExpiration())){
+                throw new SignatureException(jwtConfig.getHeader() + "失效，请重新登录。");
+            }
+        }catch (Exception e){
+            throw new SignatureException(jwtConfig.getHeader() + "失效，请重新登录。");
+        }
+        // 验证成功，根据账号返回
+        String teacher_no = claims.getSubject();
+        Teacher teacher = this.teacherDao.selectTeacherByNo(teacher_no);
+        if(teacher != null) {
+            teacher.setTeacher_password("");
+        }
+        return teacher;
     }
 
     @Override
@@ -43,8 +76,16 @@ public class TeacherServiceImpl implements TeacherService {
 
     @Override
     public boolean deleteTeacher(String teacher_no) {
-        int res = this.teacherDao.deleteTeacher(teacher_no);
-        return res > 0;
+        // 判断是否有职位，若有则不可删除
+        List<Position> positionList =  this.positionService.findPositionsByTeacher_no(teacher_no);
+        if(positionList == null || positionList.size() == 0){
+            int res = this.teacherDao.deleteTeacher(teacher_no);
+            return res > 0;
+        }else{
+            // 有职位不可删除
+            return false;
+        }
+
     }
 
     // 避免脏读
@@ -71,5 +112,13 @@ public class TeacherServiceImpl implements TeacherService {
             teacher.setTeacher_password("");
         }
         return teacherList;
+    }
+
+    @Override
+    public Teacher getTeacherInProcedure(int pro_no, int order) {
+        TeacherQueryAssist teacherQueryAssist = new TeacherQueryAssist();
+        teacherQueryAssist.setOrder(order);
+        teacherQueryAssist.setPro_no(pro_no);
+        return this.teacherDao.getTeacherInProcedure(teacherQueryAssist);
     }
 }
